job_asb/requirements.txt
azure-servicebus
azure-identity
python-dotenv
